chore: move broadcast_protocol to t's own file
